refactor(codegen): extract elementByProp into plugin folder
