+++ scripts/codegen/src/core/registry.ts
@@ -3,6 +3,7 @@ import { branchesPlugin } from "../plugins/branches/index.ts";
 import { coveragePlugin } from "../plugins/coverage/index.ts";
 import { defaultChildrenPlugin } from "../plugins/defaultChildren/index.ts";
 import { dependenciesPlugin } from "../plugins/dependencies/index.ts";
+import { elementByPropPlugin } from "../plugins/elementByProp/index.ts";
 import { examplesPlugin } from "../plugins/examples/index.ts";
 import { formControlPlugin } from "../plugins/formControl/index.ts";
 import { htmlAttrsPlugin } from "../plugins/htmlAttrs/index.ts";
@@ -38,4 +39,5 @@ export const PLUGINS: readonly SubstratePlugin[] = [
   htmlAttrsPlugin,
   defaultChildrenPlugin,
   polymorphicPlugin,
+  elementByPropPlugin,
 ];
