+++ scripts/codegen/src/plugins/elementByProp/analyze.ts
@@ -1,0 +1,7 @@
+import type { SpecAnalysis } from "../../core/analysis.ts";
+import type { Spec } from "../../schema.ts";
+
+export function analyzeElementByProp(spec: Spec): Partial<SpecAnalysis> {
+  if (spec.kind !== "atomic") return {};
+  return { elementByPropControllingProp: spec.elementByProp?.prop };
+}
