+++ scripts/codegen/src/plugins/elementByProp/schema.ts
@@ -1,0 +1,6 @@
+import { z } from "zod";
+
+export const elementByPropBlock = z.strictObject({
+  prop: z.string().min(1),
+  map: z.record(z.string().min(1), z.string().min(1)),
+});
