+++ scripts/codegen/src/semantic-checks.ts
@@ -10,6 +10,7 @@ import type { Vocabulary } from "./lib/vocabulary.ts";
 import { checkA11yRefs } from "./plugins/a11y/check.ts";
 import { checkBranches } from "./plugins/branches/check.ts";
 import { checkCoverageShape } from "./plugins/coverage/check.ts";
+import { checkElementByProp } from "./plugins/elementByProp/check.ts";
 import { checkExamplesPresent } from "./plugins/examples/check.ts";
 import { checkFormControl } from "./plugins/formControl/check.ts";
 import { checkImperativeProps } from "./plugins/imperativeProps/check.ts";
@@ -348,82 +349,6 @@ function visitPart(
   for (const [childName, child] of Object.entries(part.parts ?? {})) {
     visitPart(child, `${path}.parts.${childName}`, visit);
   }
-}
-
-// ── `elementByProp` constraints ────────────────────────────────────────────
-
-/**
- * `elementByProp` resolves the rendered tag at runtime from a controlling
- * prop's value. Four rules keep the generator output sound:
- *
- * - Mutually exclusive with sibling `element` (the tag has one source).
- * - The named prop must exist on the same node.
- * - The named prop must be `type: 'string'` — boolean / number / responsive
- *   prop types are out of scope for the v1 surface.
- * - The named prop's `values:` must enumerate exactly the map's keys (any
- *   value the consumer can pass must hit the map; every map key must be a
- *   valid prop value).
- *
- * Walks atomic root + composite parts via `visitNodes`.
- */
-export function checkElementByProp(spec: Spec): Issue[] {
-  const issues: Issue[] = [];
-  visitNodes(spec, (node, path) => {
-    const ebp = node.elementByProp;
-    if (!ebp) return;
-    const base = path === "" ? "elementByProp" : `${path}.elementByProp`;
-    if (node.element) {
-      issues.push(
-        issue(
-          spec.name,
-          base,
-          "`elementByProp` and `element` are mutually exclusive — the rendered tag has one source",
-        ),
-      );
-    }
-    const controllingProp = node.props?.[ebp.prop];
-    if (!controllingProp) {
-      issues.push(
-        issue(spec.name, `${base}.prop`, `prop '${ebp.prop}' is not declared on this node`),
-      );
-      return;
-    }
-    if (controllingProp.type !== "string") {
-      issues.push(
-        issue(
-          spec.name,
-          `${base}.prop`,
-          `prop '${ebp.prop}' must be \`type: 'string'\`; got '${controllingProp.type}'`,
-        ),
-      );
-    }
-    const mapKeys = Object.keys(ebp.map).sort();
-    const values = (controllingProp.values ?? []).slice().sort();
-    if (mapKeys.length === 0) {
-      issues.push(issue(spec.name, `${base}.map`, "`map` must declare at least one entry"));
-    }
-    const missingFromMap = values.filter((v) => !Object.hasOwn(ebp.map, v));
-    const missingFromValues = mapKeys.filter((k) => !values.includes(k));
-    if (missingFromMap.length > 0) {
-      issues.push(
-        issue(
-          spec.name,
-          `${base}.map`,
-          `prop value(s) [${missingFromMap.join(", ")}] are not in the map`,
-        ),
-      );
-    }
-    if (missingFromValues.length > 0) {
-      issues.push(
-        issue(
-          spec.name,
-          `${base}.map`,
-          `map key(s) [${missingFromValues.join(", ")}] are not declared in \`${ebp.prop}.values\``,
-        ),
-      );
-    }
-  });
-  return issues;
 }
 
 export { checkA11yRefs };
@@ -1749,6 +1674,7 @@ export { checkBranches } from "./plugins/branches/check.ts";
 export { checkCoverageShape } from "./plugins/coverage/check.ts";
 export type { DependencyIndex } from "./plugins/dependencies/check.ts";
 export { checkDependencyCycles } from "./plugins/dependencies/check.ts";
+export { checkElementByProp } from "./plugins/elementByProp/check.ts";
 export { checkExamplesPresent } from "./plugins/examples/check.ts";
 export { checkFormControl } from "./plugins/formControl/check.ts";
 export { checkImperativeProps } from "./plugins/imperativeProps/check.ts";
